修复收藏歌曲和收藏歌单功能，添加歌曲验证，添加歌曲到收藏和自定义歌单会先检查数据库是否存在该歌曲

+++ src/main/java/xyz/ztzhome/zblog/entity/Bean/Song.java
@@ -54,6 +54,6 @@ public class Song {
     /**
      * 播放次数
      */
-    private int playCount;
+    private long playCount;
 
 }
